Copy sources from the Launchpad v0.3 tarball
More specifically, from there:
https://launchpad.net/indicator-multiload/stable/0.3/+download/indicator-multiload-0.3.tar.gz

Sources in the original repo did not include things like the `configure`
script, as well as C files built from Vala files, but since this
repository is not intended for the actual development, it's easier if we
include these things right here. As one of the major reasons, the Vala
sources are incompatible anymore with the present-day Vala (tried with
v0.56.16), and I don't really know how to fix this without spending too
much time.